WIP: make pindeps generate dependencies for rebuilding deps
on the first run, deps will be (almost) empty because planex-depend
will not be passed any arguments.   Make will then try to rebuild
the included makefiles again and this time planex-depend will
be passed the dependencies written by planex-pindep on the previous
round, so it will generate deps for the build.   Make will rebuild
again, but nothing will change on this round so it will start building
the packages.

Signed-off-by: Euan Harris <euan.harris@citrix.com>